Increases tx-local name/key id space of legacy index commands
Legacy index commands don't take advantage of the token system in Neo4j.
Perhaps that will change, but until then each transaction having one or
more legacy index commands will also have an IndexDefineCommand which
contains a String->id mapping where all other legacy index commands in
that transaction will refer to ids instead of strings when it comes to
names/keys. Previously this id space was only 6 bits, yielding a maximum
of 63 names and 63 keys within the same transaction.

This has turned out to be a problem in some scenarios so this commit
extends that id space to 16 bits, now ranging up to 65535 names/keys in
the same transaction. This means that the log format has changed. Although
don't panic since the following commit will deal with this.

+++ community/kernel/src/test/java/org/neo4j/kernel/impl/transaction/log/BatchingTransactionAppenderTest.java
@@ -347,7 +347,7 @@ public class BatchingTransactionAppenderTest
     {
         Collection<Command> commands = new ArrayList<>();
         IndexDefineCommand command = new IndexDefineCommand();
-        command.init( new HashMap<String,Byte>(), new HashMap<String,Byte>() );
+        command.init( new HashMap<String,Integer>(), new HashMap<String,Integer>() );
         commands.add( command );
         PhysicalTransactionRepresentation transaction = new PhysicalTransactionRepresentation( commands );
         transaction.setHeader( new byte[0], 0, 0, 0, 0, 0, 0 );
@@ -415,8 +415,8 @@ public class BatchingTransactionAppenderTest
         {
             IndexDefineCommand defineCommand = new IndexDefineCommand();
             defineCommand.init(
-                    MapUtil.<String,Byte>genericMap( "one", (byte)1 ),
-                    MapUtil.<String,Byte>genericMap( "two", (byte)2 ) );
+                    MapUtil.<String,Integer>genericMap( "one", 1 ),
+                    MapUtil.<String,Integer>genericMap( "two", 2 ) );
             commands.add( defineCommand );
         }
         else
